huggingface/smoldocling_code/requirements.txt
transformers>=4.51.0
torch>=2.6.0
Pillow>=10.0.0
docling-core>=2.0.0
requests>=2.31.0
